Isometric example cleanup
