+++ Cargo.toml
@@ -15,7 +15,6 @@ bevy = [ "dep:bevy", "dep:seldom_fn_plugin", "dep:seldom_interop" ]
 default = [ "bevy" ]
 log = []
 state = [ "dep:seldom_state" ]
-iso_map = [ "dep:bevy_ecs_tilemap", "dep:seldom_interop" ]
 
 [dependencies]
 bevy = { version = "0.11", default-features = false, optional = true }
@@ -27,13 +26,7 @@ seldom_fn_plugin = { version = "0.4", optional = true }
 seldom_interop = { version = "0.4", optional = true }
 seldom_state = { version = "0.7", optional = true }
 
-# crates.io version is still bugged, so take it from github
-bevy_ecs_tilemap = { git = "https://github.com/StarArawn/bevy_ecs_tilemap.git", optional = true }
-
 [dev-dependencies]
 bevy = "0.11"
+bevy_ecs_tilemap = { git = "https://github.com/StarArawn/bevy_ecs_tilemap.git", rev = "fbda80c" }
 rand = "0.8"
-
-[[example]]
-name="isometric_map"
-required-features=["iso_map"]
